Fix Elven Accuracy to trigger on spells
